+++ controller/Api/InputManager/Digital/IDigitalInput.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Synthesis.Simulator.Input
+namespace SynthesisAPI.InputManager.Digital
 {
     // TODO: Maybe make the functions all into one that returns an enum of the state of the button
     public interface IDigitalInput
