--- engine/CoreEngine/Assets/Scripts/Engine/Input/Axis/DeadAxis.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2c2f619d36ab4dc40af008da09b26b00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
