--- engine/CoreEngine/Assets/Scripts/Engine/Input/Axis/DualAxis.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ad52165b806e72a49a15e50fe754db04
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
