--- engine/CoreEngine/Assets/Scripts/Engine/Input/Axis/IAxisInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a492bfb9f54966459d678b725a607ad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
