--- engine/CoreEngine/Assets/Scripts/Engine/Input/Axis/JoystickAxis.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2dccc220a564dd84baa4891728b43a3f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
