--- engine/CoreEngine/Assets/Scripts/Engine/Input/Axis/UnityAxis.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: abdc0fe6955f69a468362a290d5ac678
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
