--- engine/CoreEngine/Assets/Scripts/Engine/Input/Digital/ButtonDigital.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f584ab80998fd8443bb5fdd191807a27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
