--- engine/CoreEngine/Assets/Scripts/Engine/Input/Digital/IDigitalInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c446aebcf05884f4a94a384f118d0a67
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
