--- engine/CoreEngine/Assets/Scripts/Engine/Input/InputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d937e2a3f99a1984da3abd37463c48cc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
