--- engine/CoreEngine/Assets/Scripts/Engine/UI/Style/PanelStyle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cf89e75c94ba67f4ba159ef05bf29c4a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
