--- engine/CoreEngine/Assets/Scripts/Engine/UI/Style/Style.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a647d377b60cd0d4f86e6bc4ec9bef0f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
