--- engine/CoreEngine/Assets/Scripts/Engine/UI/Style/StyleHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 21e8edeaa04f3d1429e6bd4851fbec4d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
