--- engine/CoreEngine/Assets/Scripts/Engine/UI/Style/StyledComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2b26a7a156218764f99a745d5e8f9ac7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
