--- engine/CoreEngine/Assets/Scripts/Engine/UI/Style/TextStyle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 258758b5947a50149855c387d225f4f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
