--- engine/CoreEngine/Assets/Scripts/Engine/Util/SynList.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e5e727283b0bb24394bdf277083163b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
