+++ engine/CoreEngine/Assets/Scripts/Testing/TestInput2.cs
@@ -1,0 +1,27 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+// using SynthesisAPI.InputManager;
+
+// This whole class will be commented out when the API isn't in the project.
+public class TestInput2 : MonoBehaviour
+{
+    //private KeyDigital k;
+
+    //// Start is called before the first frame update
+    //void Start()
+    //{
+    //    k = (KeyDigital)(KeyCode.Space);
+    //}
+
+    //// Update is called once per frame
+    //void Update()
+    //{
+    //    IDigitalInput.DigitalState state;
+    //    if ((state = k.GetState()) != IDigitalInput.DigitalState.None)
+    //    {
+    //        if (state == IDigitalInput.DigitalState.Down) Debug.Log("DOWN");
+    //        else if (state == IDigitalInput.DigitalState.Up) Debug.Log("UP");
+    //    }
+    //}
+}
