+++ engine/CoreEngine/Assets/Scripts/Testing/TestInput2.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 50e9b65539e0d4242901fe285e245395
+guid: 897e04cc007bafa4bb26fcda1dda76e6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
